+++ test/core/buffers_adapter.cpp
@@ -8,10 +8,12 @@
 // Test that header file is self-contained.
 #include <beast/core/buffers_adapter.hpp>
 
+#include <beast/core/ostream.hpp>
 #include <beast/core/streambuf.hpp>
 #include <beast/unit_test/suite.hpp>
 #include <boost/asio/buffer.hpp>
 #include <boost/asio/streambuf.hpp>
+#include <boost/lexical_cast.hpp>
 #include <iterator>
 
 namespace beast {
@@ -24,14 +26,8 @@ public:
     std::string
     to_string(ConstBufferSequence const& bs)
     {
-        using boost::asio::buffer_cast;
-        using boost::asio::buffer_size;
-        std::string s;
-        s.reserve(buffer_size(bs));
-        for(auto const& b : bs)
-            s.append(buffer_cast<char const*>(b),
-                buffer_size(b));
-        return s;
+        return boost::lexical_cast<
+            std::string>(buffers(bs));
     }
 
     void testBuffersAdapter()