+++ CHANGELOG.md
@@ -9,6 +9,7 @@ API Changes:
 
 * Refactor http::header contents
 * New ostream() returns dynamic buffer output stream
+* New buffers() replaces to_string()
 
 --------------------------------------------------------------------------------